base collection introduced
